backend/requirements.txt
flask
pytest
psycopg2-binary
python-dotenv

